mailrs
Module slint_generatedAppWindow

Structs

AppWindow

InnerAppWindow

InnerAppWindowFieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerAppWindow

InnerButton_root_5

InnerButton_root_5FieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerButton_root_5

InnerColorSchemeSelector_18 🔒

InnerColorSchemeSelector_18FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerColorSchemeSelector_18

InnerComponent_image_9 🔒

InnerComponent_image_9FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerComponent_image_9

InnerComponent_rectangle_15 🔒

InnerComponent_rectangle_15FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerComponent_rectangle_15

InnerComponent_text_11 🔒

InnerComponent_text_11FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerComponent_text_11

InnerFluentPalette_19 🔒

InnerFluentPalette_19FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerFluentPalette_19

SharedGlobals 🔒

TextStyle

Constants

_THE_SAME_VERSION_MUST_BE_USED_FOR_THE_COMPILER_AND_THE_RUNTIME 🔒