mailrs
Module slint_generatedAppWindow

Structs

AppWindow

Durations

Facade

Font

InnerAppWindow

InnerAppWindowFieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerAppWindow

InnerAutoCompleteList_root_47

InnerAutoCompleteList_root_47FieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerAutoCompleteList_root_47

InnerColorSchemeSelector_70 🔒

InnerColorSchemeSelector_70FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerColorSchemeSelector_70

InnerComponent_rectangle_19 🔒

InnerComponent_rectangle_36 🔒

InnerComponent_rectangle_52 🔒

InnerComponent_rectangle_19FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerComponent_rectangle_19

InnerComponent_rectangle_36FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerComponent_rectangle_36

InnerComponent_rectangle_52FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerComponent_rectangle_52

InnerFacade 🔒

InnerFacadeFieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerFacade

InnerFluentPalette_71 🔒

InnerFluentPalette_71FieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerFluentPalette_71

InnerLineEditBase_root_59

InnerLineEditBase_root_59FieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerLineEditBase_root_59

InnerLineEdit_root_63

InnerLineEdit_root_63FieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerLineEdit_root_63

InnerMessageList_root_33

InnerMessageList_root_33FieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerMessageList_root_33

InnerScrollBar_root_22

InnerScrollBar_root_22FieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerScrollBar_root_22

InnerTagList_root_17

InnerTagList_root_17FieldsOffsets

Helper struct containing the offsets of the fields of the struct InnerTagList_root_17

InnerTheme 🔒

InnerThemeFieldsOffsets 🔒

Helper struct containing the offsets of the fields of the struct InnerTheme

MBox

Message

Palette

SharedGlobals 🔒

Spaces

Tag

TextStyle

Theme

Typo

Enums

CurrentFocus

Constants

_THE_SAME_VERSION_MUST_BE_USED_FOR_THE_COMPILER_AND_THE_RUNTIME 🔒

Statics

SLINT_EMBEDDED_RESOURCE_0 🔒

SLINT_EMBEDDED_RESOURCE_1 🔒

SLINT_EMBEDDED_RESOURCE_2 🔒

SLINT_EMBEDDED_RESOURCE_3 🔒